requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
dj-database-url
psycopg2-binary

# IA local (embeddings / RAG)
sentence-transformers
transformers
torch
faiss-cpu
numpy
scikit-learn
pdfminer.six
httpx
trafilatura
requests

# Test dependencies
pytest
pytest-django
